requirements.txt
torch 
numpy
subprocess
tqdm
gpytorc
botorch